requirements.txt
discord.py>=2.3
aiosqlite>=0.20
python-dotenv>=1.0
python-dateutil>=2.8
